+++ src/Features/SynologyPhotos/ClassSynologyPhotos.py
@@ -2522,6 +2522,8 @@ class ClassSynologyPhotos:
                 album_name = album.get("albumName", "")
                 if match_pattern(album_name, pattern):
                     new_name = replace_pattern(album_name, pattern=pattern, pattern_to_replace=pattern_to_replace)
+                    if not new_name or new_name == album_name:
+                        continue
                     albums_to_rename[album_id] = {
                         "album_name": album_name,
                         "new_name": new_name,